Removed the SQL format, code beautify, and minor text changes
The SQL format requires a method of passing the table name to
alasql to be successful. This would require a lower level call
to alasql than is currently implemented in the ala file in/out
nodes. Is being depreciated for now.

Code was beautified and ES6 Double Arrow functions implemented.
 (Note Arrow functions were not implemented in the
 <script></script> of node-red-contrib-ala-file.html to ensure
 compatible with older-browsers).