Update Test Flows and made starting test system same for all OS's
Changed file in/out nodes to display green status instead of grey.

Reorganized the Test flows, added a delay between the output tests
and input test to give time for all the outputs to be complete
before reading them back in.

Changed the `npm test` procedure to be identical for Linux/OSX and
Windows - only differnce being *test/startup.sh* or *test/statup.bat*

Continued the never ending documentation changes to the test
sub-directory readme.md and comments/nodes in the test flows.